ingestion/exploration_tests/finnhub-websocket-producer/requirements.txt
azure-eventhub
finnhub-python==2.4.27
websocket-client==1.9.0
websockets==15.0.1